Per-layout user statistics and locking based on queries.
Users now have per-layout everything.

User locked/beaten info is now stored based on conglomerate query, which
is more correct anyway. What we really want to know is whether
proficiency has been achieved along all relevant dimensions, not what
the name of the level happens to be.

Keyboard layout can explode queries into tiny chunks as well as
simplifying them.

Controller uses queries instead of paths to determine locked/unlocked
and beaten/not-beaten.

Sibling navigation is now available on groups and levels (allowing
beating one level to unlock the next). Nothing uses this yet, and it is
incomplete. It will need to be enhanced to handle beating a high-level
group and only unlocking the low-level first sibling of the next group.