Fixed bugs related to layout name in user stats.
The transition is now basically complete. User stats are now partitioned
based on keyboard layout so a user can learn a new layout without
overwriting existing data for an old one (or inheriting it, which is
also supremely unhelpful). Whether a level is unlocked or beaten is
entirely encompassed in a keyboard layout query, which saves space and
simplifies the data structures.