Make beaten work, compute unlockedness.
Whether something is unlocked is based on what has been so far beaten,
instead of pushing, we pull the data from a tree computation. It's not
the most efficient (could stand to be memoized), but it works.

Fixed a bug with User deserialization.

Added beaten/unlocked classes to groups and supergroups.